MscmpSystAuthn.Types.AuthenticatorResult (mscmp_syst_authn v0.1.0)

Represents the returned data values that were generated by the Authenticator creation process.

The included values and attributes will vary somewhat from Authenticator type to Authenticator type, but they will all follow this basic form.

Important!

Some Authenticator types, such as Validation and Recovery, provide values using this struct which are both security sensitive and not recoverable once discarded.

Types

t()

@type t() :: %MscmpSystAuthn.Types.AuthenticatorResult{
  access_account_id: MscmpSystAuthn.Types.access_account_id(),
  account_identifier: MscmpSystAuthn.Types.account_identifier(),
  credential: MscmpSystAuthn.Types.credential(),
  validation_credential: MscmpSystAuthn.Types.credential_id() | nil,
  validation_identifier: MscmpSystAuthn.Types.account_identifier() | nil
}

Represents the returned data values that were generated by the Authenticator creation process.

Attributes

access_account_id - the Access Account record ID for which the Authenticator has been created.

account_identifier - the Access Account Identifier for use in identifying the Access Account in the authentication process. Depending on the Authenticator type, this could be a restatement of what the user provided, such as an Email address, or may be system generated such as the identifier for user with an API Token.

credential - the plaintext Credential created for the Authenticator. This attribute is only returned in cases where the typical Authenticator creation flow includes the system randomly generating the Credential. For example, an API Token Credential is automatically created by the system and the Credential must be communicated to the user once created. Note that even in these cases the only time the Credential is available for disclosure to the user is immediately after the Authenticator creation via this value. Once the return result is discarded the Credential is unrecoverable.

validation_identifier - when creating an Email/Password Authenticator, the default options mandate that validating the Email Identity is required. Assuming the default is not overridden, the system generated Identifier for the user to use to validate the Email Identity is provided via this attribute. If validation is not required this attribute will not be present in the return value of the Authenticator creation call.

validation_credential - the (usually) system generated Credential in plaintext which compliments the validation_identifier value. This attribute is either included or excluded on the same terms as the validation_identifier. Note that the return of this value after the creation of an Email/Password Authenticator is the only time that this value is available in plaintext; after this point the plaintext is not retrievable.

See MscmpSystAuthn.Types.AuthenticatorResult for more.