Msdata.SystAccessAccountInstanceAssocs (mscmp_syst_authn v0.1.0)

Associates access accounts with the instances for which they are allowed to authenticate to.

Note that being able to authenticate to an instance is not the same as having authorized rights within the instance; authorization is handled by the instance directly.

Defined in MscmpSystAuthn.

Types

t()

@type t() :: %Msdata.SystAccessAccountInstanceAssocs{
  __meta__: Ecto.Schema.Metadata.t(),
  access_account:
    Msdata.SystAccessAccounts.t() | Ecto.Association.NotLoaded.t() | nil,
  access_account_id: MscmpSystAuthn.Types.access_account_id() | nil,
  access_granted: DateTime.t() | nil,
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  id: Ecto.UUID.t() | nil,
  instance: Msdata.SystInstances.t() | Ecto.Association.NotLoaded.t() | nil,
  instance_id: MscmpSystInstance.Types.instance_id() | nil,
  invitation_declined: DateTime.t() | nil,
  invitation_expires: DateTime.t() | nil,
  invitation_issued: DateTime.t() | nil
}

Functions

insert_changeset(insert_params)

@spec insert_changeset(MscmpSystAuthn.Types.access_account_instance_assoc_params()) ::
  Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystAccessAccountInstanceAssocs.Validators.insert_changeset/1.

update_changeset(access_account_instance_assoc, update_params)

@spec update_changeset(
  t(),
  MscmpSystAuthn.Types.access_account_instance_assoc_params()
) :: Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystAccessAccountInstanceAssocs.Validators.update_changeset/2.