Msdata.SystAccessAccounts (mscmp_syst_authn v0.1.0) (mscmp_syst_authn)

Contains the known login accounts which are used solely for the purpose of authentication of users.

Authorization is handled on a per-instance basis within the application.

Defined in MscmpSystAuthn

Types

t()

@type t() :: %Msdata.SystAccessAccounts{
  __meta__: Ecto.Schema.Metadata.t(),
  access_account_instance_assocs: term(),
  access_account_state:
    Msdata.SystEnumItems.t() | Ecto.Association.NotLoaded.t() | nil,
  access_account_state_id: MscmpSystAuthn.Types.access_account_state_id() | nil,
  allow_global_logins: boolean() | nil,
  credentials: term(),
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  external_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  identities: term(),
  internal_name: MscmpSystAuthn.Types.access_account_name() | nil,
  owning_owner: Msdata.SystOwners.t() | Ecto.Association.NotLoaded.t() | nil,
  owning_owner_id: MscmpSystInstance.Types.owner_id() | nil,
  password_history: term()
}

Functions

insert_changeset(insert_params, opts \\ [])

@spec insert_changeset(MscmpSystAuthn.Types.access_account_params(), Keyword.t()) ::
  Ecto.Changeset.t()

Produces a changeset for inserting a new Access Account.

Parameters

insert_params - The parameters to be used to create the new Access Account.

opts - A keyword list of options which can customize the changeset creation process.

Options

:min_internal_name_length (pos_integer/0) - Sets the minimum grapheme length of internal_name values. The default value is 6.

:max_internal_name_length (pos_integer/0) - Sets the maximum grapheme length of internal_name values. The default value is 64.

:min_display_name_length (pos_integer/0) - Sets the minimum grapheme length of display_name values. The default value is 6.

:max_display_name_length (pos_integer/0) - Sets the maximum grapheme length of display_name values. The default value is 64.

update_changeset(access_account, update_params, opts \\ [])

@spec update_changeset(
  t(),
  MscmpSystAuthn.Types.access_account_params(),
  Keyword.t()
) :: Ecto.Changeset.t()

Produces a changeset for inserting a new Access Account.

Parameters

access_account - The Access Account to be updated.

update_params - The parameters to be used to update the Access Account.

opts - A keyword list of options which can customize the changeset update process.

Options

:min_internal_name_length (pos_integer/0) - Sets the minimum grapheme length of internal_name values. The default value is 6.

:max_internal_name_length (pos_integer/0) - Sets the maximum grapheme length of internal_name values. The default value is 64.

:min_external_name_length (pos_integer/0) - Sets the minimum grapheme length of external_name values. The default value is 2.

:max_external_name_length (pos_integer/0) - Sets the maximum grapheme length of external_name values. The default value is 64.