Msdata.SystDisallowedHosts (mscmp_syst_authn v0.1.0)

A simple listing of "banned" IP address which are not allowed to authenticate their users to the system.

This registry differs from the syst_*_network_rules tables in that IP addresses here are registered as the result of automatic system heuristics whereas the network rules are direct expressions of system administrator intent. The timing between these two mechanisms is also different in that records in this table are evaluated prior to an authentication attempt and most network rules are processed in the authentication attempt sequence.

Defined in MscmpSystAuthn.

Types

t()

@type t() :: %Msdata.SystDisallowedHosts{
  __meta__: Ecto.Schema.Metadata.t(),
  host_address: MscmpSystDb.DbTypes.Inet.t() | nil,
  id: term()
}

Functions

insert_changeset(host_address)

@spec insert_changeset(MscmpSystDb.DbTypes.Inet.t()) :: Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystDisallowedHosts.Validators.insert_changeset/1.