Msdata.SystIdentities (mscmp_syst_authn v0.1.0)

The identities with which access accounts are identified to the system.

The most common example of an identity would be a user name such as an email address.

Defined in MscmpSystAuthn.

Types

t()

@type t() :: %Msdata.SystIdentities{
  __meta__: Ecto.Schema.Metadata.t(),
  access_account:
    Msdata.SystAccessAccounts.t() | Ecto.Association.NotLoaded.t() | nil,
  access_account_id: MscmpSystAuthn.Types.access_account_id() | nil,
  account_identifier: String.t() | nil,
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  external_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  identity_expires: DateTime.t() | nil,
  identity_type:
    Msdata.SystEnumItems.t() | Ecto.Association.NotLoaded.t() | nil,
  identity_type_id: MscmpSystAuthn.Types.identity_type_id() | nil,
  validated: DateTime.t() | nil,
  validates_identity: t() | Ecto.Association.NotLoaded.t() | nil,
  validates_identity_id: MscmpSystAuthn.Types.identity_id() | nil,
  validation_requested: DateTime.t() | nil
}

Functions

insert_changeset(insert_params)

@spec insert_changeset(MscmpSystAuthn.Types.identity_params()) :: Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystIdentities.Validators.insert_changeset/1.

update_changeset(identity, update_params)

@spec update_changeset(t(), MscmpSystAuthn.Types.identity_params()) ::
  Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystIdentities.Validators.update_changeset/2.