Msdata.SystOwnerNetworkRules (mscmp_syst_authn v0.1.0)

Expresses Owner specific rules concerning which hosts, as identified by IP address, may or may not attempt to authenticate with the system.

These are part of a firewall-like set of rules of which those defined in this 'owner' scope are evaluated after any SystGlobalNetworkRules defined rules and prior to any SystInstanceNetworkRules defined rules.

Defined in MscmpSystAuthn.

Types

t()

@type t() :: %Msdata.SystOwnerNetworkRules{
  __meta__: Ecto.Schema.Metadata.t(),
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  functional_type: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  ip_family: integer() | nil,
  ip_host_or_network: MscmpSystDb.DbTypes.Inet.t() | nil,
  ip_host_range_lower: MscmpSystDb.DbTypes.Inet.t() | nil,
  ip_host_range_upper: MscmpSystDb.DbTypes.Inet.t() | nil,
  ordering: integer() | nil,
  owner: Msdata.SystOwners.t() | Ecto.Association.NotLoaded.t() | nil,
  owner_id: MscmpSystInstance.Types.owner_id() | nil
}

Functions

insert_changeset(insert_params)

@spec insert_changeset(MscmpSystAuthn.Types.owner_network_rule_params()) ::
  Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystOwnerNetworkRules.Validators.insert_changeset/1.

update_changeset(owner_network_rule, update_params)

@spec update_changeset(t(), MscmpSystAuthn.Types.owner_network_rule_params()) ::
  Ecto.Changeset.t()

See MscmpSystAuthn.Impl.Msdata.SystOwnerNetworkRules.Validators.update_changeset/2.