MscmpSystAuthn.Types.AppliedNetworkRule (mscmp_syst_authn v0.1.0)

Represents the return value of functions which evaluate whether or not a specific originating host IP address is allowed to attempt an authentication.

When an authentication is attempted, a number of different rules and data sources may determine what kind of connection is allowed based on the host IP address from which the attempt appears to be originating and the destination Instance for which the authentication is being processed. To normalize the return from the evaluation of these different network based evaluations we define a simplified t:applied_network_rule/0 value to represent the actionable data of the evaluation.

Types

t()

@type t() :: %MscmpSystAuthn.Types.AppliedNetworkRule{
  functional_type: MscmpSystAuthn.Types.network_rule_functional_type(),
  network_rule_id: Ecto.UUID.t() | nil,
  precedence: MscmpSystAuthn.Types.network_rule_precedence()
}

Represents the return value of functions which evaluate whether or not a specific originating host IP address is allowed to attempt an authentication.

Attributes

precedence - this will indicate the precedence of the applied Network Rule. This value can aid in identifying the origin of a given applied allowance or denial.

network_rule_id - this is the record ID of the rule which was evaluated to govern the authentication attempt. Which specific relation the ID refers to will depend on the specific precedence of the rule applied. This value may be nil if the implied default Network Rule is applied since, by definition, this means no other explicitly defined Network Rule was found.

functional_type - May be either :allow or :deny. :allow means that the Network Rule evaluation allows the host IP address to attempt authentication. :deny indicates that the host IP address may not attempt an authentication.

See MscmpSystAuthn.Types.AppliedNetworkRule for more.